Rename mobControllerProvider file, Remove test
