+++ lib/providers/mob_state_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mob_app/providers/mob.dart';
+import 'package:mob_app/providers/mob_controller_provider.dart';
 
 class MobStateNotifier extends Notifier<MobState> {
   @override
